Add "First_provided_column" type
Like coalesce, but works on columns/column definitions that exist/are valid, rather than the contents of said columns.
Useful for data sources in which columns may be renamed at some point and you want to process files from before and after the name change or when columns are added at a point and are missing from earlier files.